+++ webapi/src/main/scala/org/knora/webapi/messages/v2/responder/KnoraResponseV2.scala
@@ -74,7 +74,7 @@ trait KnoraJsonLDResponseV2 extends KnoraResponseV2 {
       case nonJsonLD: NonJsonLD =>
         // Some other format. Convert the JSON-LD document to an RDF model.
         val rdfFormatUtil: RdfFormatUtil = RdfFeatureFactory.getRdfFormatUtil()
-        val rdfModel: RdfModel           = jsonLDDocument.toRdfModel(rdfFormatUtil.getRdfModelFactory)
+        val rdfModel: RdfModel           = jsonLDDocument.toRdfModel(rdfFormatUtil.modelFactory)
 
         // Convert the model to the requested format.
         rdfFormatUtil.format(
